requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==1.3.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==35.0.0
Faker==9.8.0
h11==0.12.0
idna==3.3
iniconfig==1.1.1
outcome==1.1.0
packaging==21.2
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
text-unidecode==1.3
toml==0.10.2
trio==0.17.0
trio-websocket==0.9.0
urllib3==1.26.2
wsproto==1.0.0
